SECOND FIX Очередь строительства

+++ my_game/building/assembly_line_workpieces.py
@@ -12,35 +12,36 @@ def check_assembly_line_workpieces(request):
     # перебор елементов очереди
     for turn_assembly_pieces in turn_assembly_piecess:
         time_start = turn_assembly_pieces.start_time_assembly
-        delta_time = time - time_start
-        new_delta = delta_time.seconds
-        delta_time = turn_assembly_pieces.finish_time_assembly - turn_assembly_pieces.start_time_assembly
-        delta = delta_time.seconds
-        if turn_assembly_pieces.class_id != 14:
-            warehouse_factory = WarehouseFactory.objects.filter(factory=turn_assembly_pieces.pattern).first()
-        else:
-            warehouse_factory = WarehouseBuilding.objects.filter(building=turn_assembly_pieces.pattern).first()
-        # проверка времени
-        if new_delta > delta:
-            if warehouse_factory is not None:
-                amount_assembly = turn_assembly_pieces.amount_assembly + warehouse_factory.amount
-                setattr(warehouse_factory, 'amount', amount_assembly)
-                warehouse_factory.save()
+        if time > time_start:
+            delta_time = time - time_start
+            new_delta = delta_time.total_seconds()
+            delta_time = turn_assembly_pieces.finish_time_assembly - turn_assembly_pieces.start_time_assembly
+            delta = delta_time.total_seconds()
+            if turn_assembly_pieces.class_id != 14:
+                warehouse_factory = WarehouseFactory.objects.filter(factory=turn_assembly_pieces.pattern).first()
             else:
-                if turn_assembly_pieces.class_id != 14:
-                    new_factory = WarehouseFactory(
-                        user=turn_assembly_pieces.user,
-                        user_city=turn_assembly_pieces.user_city,
-                        factory=turn_assembly_pieces.pattern,
-                        amount=turn_assembly_pieces.amount_assembly
-                    )
-                    new_factory.save()
+                warehouse_factory = WarehouseBuilding.objects.filter(building=turn_assembly_pieces.pattern).first()
+            # проверка времени
+            if new_delta > delta:
+                if warehouse_factory is not None:
+                    amount_assembly = turn_assembly_pieces.amount_assembly + warehouse_factory.amount
+                    setattr(warehouse_factory, 'amount', amount_assembly)
+                    warehouse_factory.save()
                 else:
-                    new_building = WarehouseBuilding(
-                        user=turn_assembly_pieces.user,
-                        user_city=turn_assembly_pieces.user_city,
-                        building=turn_assembly_pieces.pattern,
-                        amount=turn_assembly_pieces.amount_assembly
-                    )
-                    new_building.save()
-            TurnAssemblyPieces.objects.filter(id=turn_assembly_pieces.id).delete()
+                    if turn_assembly_pieces.class_id != 14:
+                        new_factory = WarehouseFactory(
+                            user=turn_assembly_pieces.user,
+                            user_city=turn_assembly_pieces.user_city,
+                            factory=turn_assembly_pieces.pattern,
+                            amount=turn_assembly_pieces.amount_assembly
+                        )
+                        new_factory.save()
+                    else:
+                        new_building = WarehouseBuilding(
+                            user=turn_assembly_pieces.user,
+                            user_city=turn_assembly_pieces.user_city,
+                            building=turn_assembly_pieces.pattern,
+                            amount=turn_assembly_pieces.amount_assembly
+                        )
+                        new_building.save()
+                TurnAssemblyPieces.objects.filter(id=turn_assembly_pieces.id).delete()
